requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==5.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.1.1
django-paypal==1.0.0
django-session-timeout==0.1.0
html5lib==1.1
idna==2.10
image==1.5.32
importlib-metadata==2.0.0
kombu==5.0.2
Pillow==7.2.0
prompt-toolkit==3.0.7
pycparser==2.20
Pyphen==0.9.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
zipp==3.2.0
